+++ Projects/UOContent.Tests/UOContent.Tests.csproj
@@ -10,5 +10,9 @@
         <ProjectReference Include="..\Server\Server.csproj" />
         <ProjectReference Include="..\Server.Tests\Server.Tests.csproj" />
         <ProjectReference Include="..\UOContent\UOContent.csproj" />
+        <DataFiles Include="$(SolutionDir)\Distribution\Data\**" />
     </ItemGroup>
+    <Target Name="CopyData" AfterTargets="AfterBuild">
+        <Copy SourceFiles="@(DataFiles)" DestinationFolder="$(OutDir)\Data\%(RecursiveDir)" />
+    </Target>
 </Project>
